requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=h0e60522_0
altair=5.1.2=pyhd8ed1ab_0
anyio=3.5.0=py311haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h2bbff1b_0
arrow-cpp=11.0.0=ha81ea56_2
asttokens=2.4.0=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
babel=2.11.0=py311haa95532_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=py311haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.3=pyhd8ed1ab_0
boost-cpp=1.78.0=h5b4e17d_0
bottleneck=1.3.5=py311h5bb9823_0
branca=0.6.0=pyhd8ed1ab_0
brotli-python=1.0.9=py311hd77b12b_7
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.08.22=haa95532_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=py311haa95532_0
cffi=1.15.1=py311h2bbff1b_3
charset-normalizer=3.3.0=pyhd8ed1ab_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
debugpy=1.6.7=py311hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py311haa95532_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
folium=0.14.0=pyhd8ed1ab_0
freetype=2.10.4=h546665d_1
gflags=2.2.2=ha925a31_1004
giflib=5.2.1=h8d14728_2
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.40=pyhd8ed1ab_0
glog=0.5.0=h4797de2_0
grpc-cpp=1.48.2=hfe90ff0_1
haversine=2.8.0=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=6.1.0=pyhd8ed1ab_0
inflection=0.5.1=py311haa95532_0
intel-openmp=2023.2.0=h57928b3_50496
ipykernel=6.25.2=pyh60829e3_0
ipython=8.16.1=pyh5737063_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.1.1=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h8ffe710_2
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=8.4.0=pyhd8ed1ab_0
jupyter_core=5.4.0=py311h1ea47a8_0
jupyter_server=1.23.4=py311haa95532_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.22.0=py311haa95532_0
jupyterlab_widgets=3.0.9=pyhd8ed1ab_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcurl=8.1.1=h86230a5_0
libdeflate=1.17=h2bbff1b_1
libevent=2.1.10=h91dae50_4
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.5.1=hd77b12b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lxml=4.9.3=py311h09808a7_0
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.1=py311h2bbff1b_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=pyhd8ed1ab_0
mistune=0.8.4=py311h2bbff1b_1000
mkl=2023.2.0=h6a75c08_50496
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
nbclassic=0.5.5=py311haa95532_0
nbclient=0.5.13=py311haa95532_0
nbconvert=6.5.4=py311haa95532_0
nbformat=5.9.2=py311haa95532_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
notebook-shim=0.2.2=py311haa95532_0
numexpr=2.8.7=py311h1fcbade_0
numpy=1.26.0=py311hdab7c0b_0
numpy-base=1.26.0=py311hd01c5d8_0
openssl=3.0.11=h2bbff1b_2
orc=1.7.4=h623e30f_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py311hf62ec03_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311hd77b12b_1
pip=23.3=py311haa95532_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=3.11.0=pyhd8ed1ab_0
plotly=5.17.0=py_0
prometheus_client=0.14.1=py311haa95532_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=3.20.3=py311h1ea47a8_1
psutil=5.9.0=py311h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py311h8a3a540_1
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyrsistent=0.18.0=py311h2bbff1b_0
pysocks=1.7.1=pyh0701188_6
python=3.11.5=he1021f5_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=py311haa95532_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytz=2023.3.post1=pyhd8ed1ab_0
pywin32=305=py311h2bbff1b_0
pywinpty=2.0.10=py311h5da7b33_0
pyyaml=6.0=py311h2bbff1b_1
pyzmq=25.1.0=py311hd77b12b_0
re2=2022.04.01=h0e60522_0
requests=2.31.0=pyhd8ed1ab_0
rich=13.6.0=pyhd8ed1ab_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=68.0.0=py311haa95532_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
snappy=1.1.9=h6c2663c_0
sniffio=1.2.0=py311haa95532_1
soupsieve=2.5=py311haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.27.2=pyhd8ed1ab_1
streamlit-folium=0.15.0=pyhd8ed1ab_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=pyhd8ed1ab_0
terminado=0.17.1=py311haa95532_0
tinycss2=1.2.1=py311haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.3=py311h2bbff1b_0
traitlets=5.11.2=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
tzlocal=5.1=py311h1ea47a8_0
urllib3=2.0.7=pyhd8ed1ab_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.22.0=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=py311h1ea47a8_1
wcwidth=0.2.8=pyhd8ed1ab_0
webencodings=0.5.1=py311haa95532_1
websocket-client=0.58.0=py311haa95532_4
wheel=0.41.2=py311haa95532_0
widgetsnbextension=4.0.9=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
winpty=0.4.3=4
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
